requirements.txt
alembic==1.7.4
anyio==3.3.3
apache-airflow==2.2.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-sqlite==2.0.1
apispec==3.3.2
argcomplete==1.12.3
astroid==2.8.4
attrs==20.3.0
Babel==2.9.1
blinker==1.4
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.6
charset-normalizer==2.0.6
click==7.1.2
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.8
defusedxml==0.7.1
dill==0.3.1.1
dnspython==2.1.0
docutils==0.16
email-validator==1.1.3
Flask==1.1.4
Flask-AppBuilder==3.3.3
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
graphviz==0.17
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.2
importlib-metadata==4.8.1
importlib-resources==5.2.2
inflection==0.5.1
iso8601==0.1.16
isodate==0.6.0
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.1.0
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.5
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.13.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
mccabe==0.6.1
mypy-extensions==0.4.3
nltk==3.6.5
oauthlib==3.1.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
packaging==21.0
pathspec==0.9.0
pendulum==2.1.2
platformdirs==2.4.0
prison==0.2.1
psutil==5.8.0
psycopg2-binary==2.9.1
pycparser==2.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.2
python-dotenv==0.19.1
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytzdata==2020.1
PyYAML==5.4.1
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rfc3986==1.5.0
rich==10.12.0
setproctitle==1.2.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
textblob==0.17.1
toml==0.10.2
tomli==1.2.2
tqdm==4.62.3
tweepy==4.1.0
typing-extensions==3.10.0.2
unicodecsv==0.14.1
urllib3==1.26.7
Werkzeug==1.0.1
wrapt==1.13.3
WTForms==2.3.3
zipp==3.6.0
